+++ Source/Scene/Model.js
@@ -334,6 +334,8 @@ define([
      * @param {Boolean} [options.asynchronous=true] Determines if model WebGL resource creation will be spread out over several frames or block until completion once all glTF files are loaded.
      * @param {Boolean} [options.debugShowBoundingVolume=false] For debugging only. Draws the bounding sphere for each draw command in the model.
      * @param {Boolean} [options.debugWireframe=false] For debugging only. Draws the model in wireframe.
+     * @param {Boolean} [options.castShadows=true] Determines whether this model will cast shadows when shadow mapping is enabled.
+     * @param {Boolean} [options.receiveShadows=true] Determines whether this model will receive shadows when shadow mapping is enabled.
      *
      * @exception {DeveloperError} bgltf is not a valid Binary glTF file.
      * @exception {DeveloperError} Only glTF Binary version 1 is supported.
@@ -582,6 +584,9 @@ define([
         // CESIUM_RTC extension
         this._rtcCenter = undefined;    // in world coordinates
         this._rtcCenterEye = undefined; // in eye coordinates
+
+        this._castShadows = defaultValue(options.castShadows, true);
+        this._receiveShadows = defaultValue(options.receiveShadows, true);
     }
 
     defineProperties(Model.prototype, {
@@ -819,6 +824,54 @@ define([
         },
 
         /**
+         * Determines whether the model will cast shadows when shadow mapping is enabled.
+         *
+         * @memberof Model.prototype
+         *
+         * @type {Boolean}
+         *
+         * @default true
+         */
+        castShadows : {
+            get : function() {
+                return this._castShadows;
+            },
+            set : function(value) {
+                if (value !== this._castShadows) {
+                    this._castShadows = value;
+                    var length = this._nodeCommands.length;
+                    for (var i = 0; i < length; ++i) {
+                        this._nodeCommands[i].command.castShadows = value;
+                    }
+                }
+            }
+        },
+
+        /**
+         * Determines whether the model will receive shadows when shadow mapping is enabled.
+         *
+         * @memberof Model.prototype
+         *
+         * @type {Boolean}
+         *
+         * @default true
+         */
+        receiveShadows : {
+            get : function() {
+                return this._receiveShadows;
+            },
+            set : function(value) {
+                if (value !== this._receiveShadows) {
+                    this._receiveShadows = value;
+                    var length = this._nodeCommands.length;
+                    for (var i = 0; i < length; ++i) {
+                        this._nodeCommands[i].command.receiveShadows = value;
+                    }
+                }
+            }
+        },
+
+        /**
          * Returns true if the model was transformed this frame
          *
          * @memberof Model.prototype
@@ -918,6 +971,9 @@ define([
      * @param {Boolean} [options.asynchronous=true] Determines if model WebGL resource creation will be spread out over several frames or block until completion once all glTF files are loaded.
      * @param {Boolean} [options.debugShowBoundingVolume=false] For debugging only. Draws the bounding sphere for each {@link DrawCommand} in the model.
      * @param {Boolean} [options.debugWireframe=false] For debugging only. Draws the model in wireframe.
+     * @param {Boolean} [options.castShadows=true] Determines whether this model will cast shadows when shadow mapping is enabled.
+     * @param {Boolean} [options.receiveShadows=true] Determines whether this model will receive shadows when shadow mapping is enabled.
+     *
      * @returns {Model} The newly created model.
      *
      * @exception {DeveloperError} bgltf is not a valid Binary glTF file.
@@ -1163,7 +1219,7 @@ define([
 
                 // The extension 'KHR_binary_glTF' uses a special buffer entitled just 'binary_glTF'.
                 // The 'KHR_binary_glTF' check is for backwards compatibility for the Cesium model converter
-                // circa Cesium 1.15-1.17 when the converter incorrectly used the buffer name 'KHR_binary_glTF'.
+                // circa Cesium 1.15-1.20 when the converter incorrectly used the buffer name 'KHR_binary_glTF'.
                 if ((id === 'binary_glTF') || (id === 'KHR_binary_glTF')) {
                     // Buffer is the binary glTF file itself that is already loaded
                     var loadResources = model._loadResources;
@@ -1718,9 +1774,6 @@ define([
         var vs = getShaderSource(model, shaders[program.vertexShader]);
         var fs = getShaderSource(model, shaders[program.fragmentShader]);
 
-        var drawVS = modifyShader(vs, id, model._vertexShaderLoaded);
-        var drawFS = modifyShader(fs, id, model._fragmentShaderLoaded);
-
         // Add pre-created attributes to attributeLocations
         var attributesLength = program.attributes.length;
         var precreatedAttributes = model._precreatedAttributes;
@@ -1731,6 +1784,9 @@ define([
                 }
             }
         }
+
+        var drawVS = modifyShader(vs, id, model._vertexShaderLoaded);
+        var drawFS = modifyShader(fs, id, model._fragmentShaderLoaded);
 
         model._rendererResources.programs[id] = ShaderProgram.fromCache({
             context : context,
@@ -2858,6 +2914,8 @@ define([
                     count : count,
                     offset : offset,
                     shaderProgram : rendererPrograms[technique.program],
+                    castShadows : model._castShadows,
+                    receiveShadows : model._receiveShadows,
                     uniformMap : uniformMap,
                     renderState : rs,
                     owner : owner,