+++ Source/Scene/Scene.js
@@ -35,6 +35,7 @@ define([
         '../Renderer/ComputeEngine',
         '../Renderer/Context',
         '../Renderer/ContextLimits',
+        '../Renderer/DrawCommand',
         '../Renderer/PassState',
         '../Renderer/ShaderProgram',
         '../Renderer/ShaderSource',
@@ -63,6 +64,7 @@ define([
         './SceneTransforms',
         './SceneTransitioner',
         './ScreenSpaceCameraController',
+        './ShadowMap',
         './SunPostProcess',
         './TweenCollection'
     ], function(
@@ -101,6 +103,7 @@ define([
         ComputeEngine,
         Context,
         ContextLimits,
+        DrawCommand,
         PassState,
         ShaderProgram,
         ShaderSource,
@@ -129,6 +132,7 @@ define([
         SceneTransforms,
         SceneTransitioner,
         ScreenSpaceCameraController,
+        ShadowMap,
         SunPostProcess,
         TweenCollection) {
     'use strict';
@@ -541,6 +545,17 @@ define([
          * @type {Fog}
          */
         this.fog = new Fog();
+
+        this._sunCamera = new Camera(this);
+
+        /**
+         * Render shadows in the scene.
+         * @type {ShadowMap}
+         */
+        this.shadowMap = new ShadowMap({
+            context : context,
+            lightCamera : this._sunCamera
+        });
 
         this._terrainExaggeration = defaultValue(options.terrainExaggeration, 1.0);
 
@@ -1049,6 +1064,31 @@ define([
             Matrix4.equalsEpsilon(camera0.transform, camera1.transform, epsilon);
     }
 
+    function updateDerivedCommands(scene, command) {
+        if (command._dirty) {
+            command._dirty = false;
+
+            var context = scene._context;
+            var frameState = scene.frameState;
+
+            var derivedCommands = command.derivedCommands;
+
+            var shadowMaps = frameState.shadowMaps;
+            if (shadowMaps.length > 0) {
+                derivedCommands.shadows = ShadowMap.createDerivedCommands(shadowMaps, command, context, derivedCommands.shadows);
+            }
+
+            var oit = scene._oit;
+            if (command.pass === Pass.TRANSLUCENT && defined(oit) && oit.isSupported()) {
+                if (shadowMaps.length > 0 && command.receiveShadows) {
+                    derivedCommands.oit = oit.createDerivedCommands(command.derivedCommands.shadows.receiveCommand, context, derivedCommands.oit);
+                } else {
+                    derivedCommands.oit = oit.createDerivedCommands(command, context, derivedCommands.oit);
+                }
+            }
+        }
+    }
+
     var scratchOccluderBoundingSphere = new BoundingSphere();
     var scratchOccluder;
 
@@ -1076,6 +1116,7 @@ define([
 
         var frameState = scene._frameState;
         frameState.commandList.length = 0;
+        frameState.shadowMaps.length = 0;
         frameState.mode = scene._mode;
         frameState.morphTime = scene.morphTime;
         frameState.mapProjection = scene.mapProjection;
@@ -1109,6 +1150,8 @@ define([
         if (scene.debugShowFrustums) {
             command.debugOverlappingFrustums = 0;
         }
+
+        updateDerivedCommands(scene, command);
 
         var frustumCommandsList = scene._frustumCommandsList;
         var length = frustumCommandsList.length;
@@ -1190,6 +1233,9 @@ define([
         var far = -Number.MAX_VALUE;
         var undefBV = false;
 
+        var shadowNear = Number.MAX_VALUE;
+        var shadowFar = -Number.MAX_VALUE;
+
         var occluder = (frameState.mode === SceneMode.SCENE3D) ? frameState.occluder: undefined;
         var cullingVolume = frameState.cullingVolume;
 
@@ -1225,6 +1271,15 @@ define([
                     distances = boundingVolume.computePlaneDistances(position, direction, distances);
                     near = Math.min(near, distances.start);
                     far = Math.max(far, distances.stop);
+
+                    // When moving the camera low LOD terrain tiles begin to load, whose bounding volumes
+                    // throw off the near/far fitting for the shadow map. Only update shadowNear and shadowFar
+                    // for reasonably sized bounding volumes.
+                    // TODO : handle similar case for 3D Tiles
+                    if (!((distances.start < -100.0) && (distances.stop > 100.0) && (pass === Pass.GLOBE))){
+                        shadowNear = Math.min(shadowNear, distances.start);
+                        shadowFar = Math.max(shadowFar, distances.stop);
+                    }
                 } else {
                     // Clear commands don't need a bounding volume - just add the clear to all frustums.
                     // If another command has no bounding volume, though, we need to use the camera's
@@ -1247,7 +1302,14 @@ define([
             // This will handle the case where the computed near plane is further than the user defined far plane.
             near = Math.min(Math.max(near, camera.frustum.near), camera.frustum.far);
             far = Math.max(Math.min(far, camera.frustum.far), near);
+
+            shadowNear = Math.min(Math.max(shadowNear, camera.frustum.near), camera.frustum.far);
+            shadowFar = Math.max(Math.min(shadowFar, camera.frustum.far), shadowNear);
         }
+
+        // Use the computed near and far for shadows
+        frameState.shadowNear = shadowNear;
+        frameState.shadowFar = shadowFar;
 
         // Exploit temporal coherence. If the frustums haven't changed much, use the frustums computed
         // last frame, else compute the new frustums and sort them by frustum again.
@@ -1317,13 +1379,11 @@ define([
         });
     }
 
-    function executeDebugCommand(command, scene, passState, renderState, shaderProgram) {
-        if (defined(command.shaderProgram) || defined(shaderProgram)) {
-            // Replace shader for frustum visualization
-            var sp = createDebugFragmentShaderProgram(command, scene, shaderProgram);
-            command.execute(scene.context, passState, renderState, sp);
-            sp.destroy();
-        }
+    function executeDebugCommand(command, scene, passState) {
+        var debugCommand = DrawCommand.shallowClone(command);
+        debugCommand.shaderProgram = createDebugFragmentShaderProgram(command, scene);
+        debugCommand.execute(scene.context, passState);
+        debugCommand.shaderProgram.destroy();
     }
 
     var transformFrom2D = new Matrix4(0.0, 0.0, 1.0, 0.0,
@@ -1332,15 +1392,17 @@ define([
                                       0.0, 0.0, 0.0, 1.0);
     transformFrom2D = Matrix4.inverseTransformation(transformFrom2D, transformFrom2D);
 
-    function executeCommand(command, scene, context, passState, renderState, shaderProgram, debugFramebuffer) {
+    function executeCommand(command, scene, context, passState, debugFramebuffer) {
         if ((defined(scene.debugCommandFilter)) && !scene.debugCommandFilter(command)) {
             return;
         }
 
         if (scene.debugShowCommands || scene.debugShowFrustums) {
-            executeDebugCommand(command, scene, passState, renderState, shaderProgram);
+            executeDebugCommand(command, scene, passState);
+        } else if (scene.frameState.shadowMaps.length > 0 && command.receiveShadows && defined(command.derivedCommands.shadows.receiveCommand)) {
+            command.derivedCommands.shadows.receiveCommand.execute(context, passState);
         } else {
-            command.execute(context, passState, renderState, shaderProgram);
+            command.execute(context, passState);
         }
 
         if (command.debugShowBoundingVolume && (defined(command.boundingVolume))) {
@@ -1471,6 +1533,8 @@ define([
         var context = scene.context;
         var us = context.uniformState;
 
+        us.updateCamera(camera);
+
         // Create a working frustum from the original camera frustum.
         var frustum;
         if (defined(camera.frustum.fov)) {
@@ -1561,6 +1625,7 @@ define([
             }
 
             us.updateFrustum(frustum);
+
             clearDepth.execute(context, passState);
 
             us.updatePass(Pass.GLOBE);
@@ -1654,6 +1719,108 @@ define([
         }
     }
 
+    function insertShadowCommands(scene, commandList, insertAll, shadowMap, shadowPassCommands) {
+        var shadowVolume = shadowMap.shadowMapCullingVolume;
+        var passVolumes = shadowMap.passCullingVolumes;
+
+        var isPointLight = shadowMap.isPointLight;
+        var center = shadowMap.pointLightPosition;
+        var radius = shadowMap.pointLightRadius;
+        var radiusSquared = radius * radius;
+
+        var numberOfPasses = shadowMap.numberOfPasses;
+
+        var length = commandList.length;
+        for (var i = 0; i < length; ++i) {
+            var command = commandList[i];
+            updateDerivedCommands(scene, command);
+
+            // Don't insert globe commands with the rest of the scene commands since they are handled separately
+            if (command.castShadows && (insertAll || (command.pass === Pass.OPAQUE || command.pass === Pass.TRANSLUCENT))) {
+                if (isPointLight) {
+                    if (defined(command.boundingVolume)) {
+                        var distance = command.boundingVolume.distanceSquaredTo(center);
+                        if (distance < radiusSquared) {
+                            for (var k = 0; k < numberOfPasses; ++k) {
+                                shadowPassCommands[k].push(command);
+                            }
+                        }
+                    }
+                } else {
+                    if (isVisible(command, shadowVolume)) {
+                        if (numberOfPasses <= 1) {
+                            shadowPassCommands[0].push(command);
+                        } else {
+                            var wasVisible = false;
+                            // Loop over cascades from largest to smallest
+                            for (var j = numberOfPasses - 1; j >= 0; --j) {
+                                var cascadeVolume = passVolumes[j];
+                                if (isVisible(command, cascadeVolume)) {
+                                    shadowPassCommands[j].push(command);
+                                    wasVisible = true;
+                                } else if (wasVisible) {
+                                    // If it was visible in the previous cascade but now isn't
+                                    // then there is no need to check any more cascades
+                                    break;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    function resetShadowCommands(passCommands) {
+        var length = passCommands.length;
+        for (var i = 0; i < length; ++i) {
+            passCommands[i].length = 0;
+        }
+    }
+
+    function executeShadowMapCommands(scene) {
+        var frameState = scene.frameState;
+        var shadowMaps = frameState.shadowMaps;
+        var shadowMapLength = shadowMaps.length;
+
+        if (shadowMapLength === 0 || scene.frameState.passes.picking) {
+            return;
+        }
+
+        var context = scene.context;
+        var uniformState = context.uniformState;
+
+        for (var i = 0; i < shadowMapLength; ++i) {
+            var shadowMap = shadowMaps[i];
+            if (shadowMap.outOfView) {
+                continue;
+            }
+
+            var shadowPassCommands = shadowMap.passCommands;
+            resetShadowCommands(shadowPassCommands);
+
+            // Insert the scene commands into the shadow map passes
+            var sceneCommands = scene.frameState.commandList;
+            insertShadowCommands(scene, sceneCommands, false, shadowMap, shadowPassCommands);
+            var terrainCommands = shadowMap.commandList;
+            insertShadowCommands(scene, terrainCommands, true, shadowMap, shadowPassCommands);
+
+            var numberOfPasses = shadowMap.numberOfPasses;
+
+            for (var j = 0; j < numberOfPasses; ++j) {
+                uniformState.updateCamera(shadowMap.passCameras[j]);
+                var passState = shadowMap.passStates[j];
+                shadowMap.updatePass(context, j);
+                var passCommands = shadowPassCommands[j];
+                var numberOfCommands = passCommands.length;
+                for (var k = 0; k < numberOfCommands; ++k) {
+                    var command = passCommands[k];
+                    executeCommand(command.derivedCommands.shadows.castCommands[i], scene, context, passState);
+                }
+            }
+        }
+    }
+
     function updateAndExecuteCommands(scene, passState, backgroundColor, picking) {
         var context = scene._context;
 
@@ -1668,6 +1835,7 @@ define([
             createPotentiallyVisibleSet(scene);
             updateAndClearFramebuffers(scene, passState, backgroundColor, picking);
             executeComputeCommands(scene);
+            executeShadowMapCommands(scene);
 
             // Based on Calculating Stereo pairs by Paul Bourke
             // http://paulbourke.net/stereographics/stereorender/
@@ -1815,6 +1983,7 @@ define([
         if (firstViewport) {
             updateAndClearFramebuffers(scene, passState, backgroundColor, picking);
             executeComputeCommands(scene);
+            executeShadowMapCommands(scene);
         }
 
         executeCommands(scene, passState);
@@ -1846,12 +2015,40 @@ define([
     function updatePrimitives(scene) {
         var frameState = scene._frameState;
 
+        scene._groundPrimitives.update(frameState);
+        scene._primitives.update(frameState);
+
+        var shadowMaps = frameState.shadowMaps;
+        var length = shadowMaps.length;
+        for (var i = 0; i < length; ++i) {
+            var shadowMap = shadowMaps[i];
+            shadowMap.update(frameState);
+
+            if (!shadowMap.outOfView && defined(scene._globe) && scene._globe.castShadows) {
+                var sceneCamera = frameState.camera;
+                var sceneCullingVolume = frameState.cullingVolume;
+                var sceneCommandList = frameState.commandList;
+
+                // Update frame state to render from the light camera
+                frameState.camera = shadowMap.shadowMapCamera;
+                frameState.cullingVolume = shadowMap.shadowMapCullingVolume;
+                frameState.commandList = shadowMap.commandList;
+
+                frameState.commandList.length = 0;
+
+                // Collect terrain commands from the light's POV
+                scene._globe.update(frameState);
+
+                // Revert back to original frame state
+                frameState.camera = sceneCamera;
+                frameState.cullingVolume = sceneCullingVolume;
+                frameState.commandList = sceneCommandList;
+            }
+        }
+
         if (scene._globe) {
             scene._globe.update(frameState);
         }
-
-        scene._groundPrimitives.update(frameState);
-        scene._primitives.update(frameState);
     }
 
     function updateAndClearFramebuffers(scene, passState, clearColor, picking) {
@@ -2032,6 +2229,13 @@ define([
         scene.fog.update(frameState);
 
         us.update(frameState);
+
+        var shadowMap = scene.shadowMap;
+        if (shadowMap.enabled) {
+            // Update the sun's direction
+            Cartesian3.negate(us.sunDirectionWC, scene._sunCamera.direction);
+            frameState.shadowMaps.push(shadowMap);
+        }
 
         scene._computeCommandList.length = 0;
         scene._overlayCommandList.length = 0;